requirements.txt
# Core Dependencies
flask==3.0.0
transformers==4.47.1
torch==2.5.1
datasets==3.2.0
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.2

# Language Detection
langdetect==1.0.9

# Utilities
tqdm==4.67.1
accelerate==1.3.0

# OCR for Screenshot Analysis
Pillow==10.4.0
pytesseract==0.3.10

# URL Scanning
requests==2.32.3
python-whois==0.9.4
